+++ src/main/java/com/challenge/helpers/PlanetsTriangle.java
@@ -22,9 +22,8 @@ public class PlanetsTriangle {
         return (trianglePath.contains(new Point2D.Double(0, 0)));
     }
 
-    public static Double calculateArea(List<Position> positions) {
+    public static Double calculatePerimeter(List<Position> positions) {
         if (!PositionValidator.ValidatePositionsSize(positions)) throw new PlanetsOutOfBoundsException(Messages.OUT_OF_BOUNDS_PLANETS);
-        // TODO: use Intstream
         double distance1 = Point2D.Double.distanceSq(positions.get(0).getCoordinateX(), positions.get(0).getCoordinateY(), positions.get(1).getCoordinateX(), positions.get(1).getCoordinateY());
         double distance2 = Point2D.Double.distanceSq(positions.get(1).getCoordinateX(), positions.get(1).getCoordinateY(), positions.get(2).getCoordinateX(), positions.get(2).getCoordinateY());
         double distance3 = Point2D.Double.distanceSq(positions.get(2).getCoordinateX(), positions.get(2).getCoordinateY(), positions.get(0).getCoordinateX(), positions.get(0).getCoordinateY());
