+++ src/main/java/com/challenge/model/enums/Weather.java
@@ -1,7 +1,7 @@
 package com.challenge.model.enums;
 
 public enum Weather {
-	NO_INFORMATION("No information"), DROUGHT("Droughness"), OPTIMAL("Optimal Conditions"), RAIN("Raining Conditions"), MAX_RAIN("Maximum Raining Conditions");
+	NO_INFORMATION("No information"), DROUGHT("Drought"), OPTIMAL("Optimal Conditions"), RAIN("Raining Conditions"), MAX_RAIN("Maximum Raining Conditions");
 
 	Weather(String s) {
 	}
