+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=password
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+forecast.amount.days=3600
